codec2: Fix dynamic resolution change in codec2 decoder
Issue: Video playback stalls on resolution change

Solution: On resolution change, during set format in decoder, check
for change in width and height and drain the decoder. Added a flag
to specify component drain with or without EOS. Component stop is
required only in case of format change. For runtime resolution and
framerate change component stop is not required.

Change-Id: Icbff300586dba2db23c7a304aa18a59181d5f10b