codec2: Dup fd on importing linear dma buffer
Issue: C2DmaBufAllocation takes ownership of the handle on importing a
buffer and closes the fd when the object is destroyed. This is preventing
the buffer from being reused in a bufferpool.

Solution: Duplicate the fd before importing the buffer in C2DmaBufAllocation
This will only close the duplicate fd. Buffer ownership will remain with the
allocator

Change-Id: I91f5d30983e25a8037b89ae61f25ab51c04bd254