metamux: Fix for corruption when old buffer is sent after caps changed
Issue: The issue happens when few buffers are acumulated in
the main pad queue and waiting for buffer on the data pad.
At this time new stream caps are received.
We are accepting the caps and the next plugins are reconfigured
with the new caps as well. When a new buffer is received on
the data pad it will be paired with an old buffer on the main pad
and will be sent to next plugin.

Fix: Wait for pending buffers to be processed before sending new caps

Change-Id: I4db1137491e3e118f9f3bcdd959a4969555706cb