camreproc: fix a memory leak issue
issue: there are addresses were not freed in ptr_array cause the
memory leak during using the IPE offline plugin.

Analysis: the function g_ptr_array_free will not release the address
that store the gstbuffer pointer when the free_segment parameter set
to FALSE. It cause the memory leak.

Fix: set the free_segment to TRUE to free the ptr_array.

Change-Id: Iea494bc99b9c42c497b05a6f5a2b901d37d3ade9
Signed-off-by: Hui Liu <quic_huliu@quicinc.com>